Logs: Show newest to oldest, group by week after 2 weeks, simplify description element. Entries: fix bug in deleting entry, add default role for translation, show error on no roll for proxy. Removed unused parameter for GroupUserItems.

+++ Happy Reader/View/Tabs/EntriesTab.xaml.cs
@@ -6,12 +6,11 @@ namespace Happy_Reader.View.Tabs
 {
 	public partial class EntriesTab : UserControl
 	{
-		private EntriesTabViewModel ViewModel { get; }
+		private EntriesTabViewModel ViewModel => (EntriesTabViewModel)DataContext;
 
 		public EntriesTab()
 		{
 			InitializeComponent();
-			ViewModel = (EntriesTabViewModel)DataContext;
 		}
 
 		private void AddEntry_Click(object sender, RoutedEventArgs e)
@@ -21,14 +20,8 @@ namespace Happy_Reader.View.Tabs
 		{
 			var button = (Button)sender;
 			var item = (DisplayEntry)button.DataContext;
-			if (item.DeletePrimed)
-			{
-				ViewModel.DeleteEntry(item);
-			}
-			else
-			{
-				item.PrimeDeletion(button);
-			}
+			if (item.DeletePrimed) ViewModel.DeleteEntry(item);
+			else item.PrimeDeletion(button);
 		}
 	}
 }
